KilimoKipya_Phase_1/backend/requirements.txt
httpx==0.27.0
fastapi==0.109.0
uvicorn[standard]
scikit-learn==1.3.1
joblib==1.3.2
pytest==8.1.1
pytest-asyncio==0.23.6
pandas==2.2.1
cryptography==42.0.5
pydantic_core==2.14.6
pymongo==4.7.1
reportlab==4.2.0
catboost==1.2.5